Simple Project, Only one commit for now
I did not do multiple commits on this project, was quite simple.
Task was from TheOdinProject(.com) and was to create a website where I displayed recipes.
I did not display the third recipe as I felt it was redundant. May add a "next recipe"
button or something like that, but other than that, this is completed.